BackEnd/requirements.txt
redis==5.0.7
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.8
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
Pillow==10.0.0
python-decouple==3.8
django-cors-headers==4.2.0
factory-boy==3.3.0
django-filter==23.2
gunicorn==21.2.0
toml==0.10.2
black==23.9.1
drf-spectacular==0.26.5
ratelimit==2.2.1
django-ratelimit==4.1.0
django-debug-toolbar==4.3.0
celery==5.4.0
passlib==1.7.4
Faker==19.6.2
coverage==7.6.10
channels==4.2.0
daphne==4.1.2
pymongo==3.12.3
mongoengine==0.27.0
channels_redis==4.2.1
python-dotenv==1.0.1
requests==2.32.3
django-channels-jwt-auth-middleware==1.0.0
bleach==6.2.0
django_cryptography==1.1